Added tests for the mapl function.
Fixed the infinite loop problem in maplist.  I had changed the
implementation of the compiler macro from using the loop keyword `on'
to using (loop for sublist = list then (cdr list) ...) but I had
forgotten to add a termination condition.  By adding
`until (endp sublist)' this fixed the infinite loop problem, and also
signals a type error if the list is not proper.

Modified the implementation of mapl to do what maplist does.