Added tests for the member function.
Fixed bug in the member function that were found as a result of the
tests (you can't use loop on because it will use atom as a test
for the end of the list).

Implemented the member-if and member-if-not functions.

Added tests for the member-if and member-if-not functions.

Updated status.text to reflect the modifications.