Moved the definition of mapcar so that it is defined when used in macroexpanders for push and pop.  Also wrapped it in eval-when. Thanks to _3b on #lisp for spotting this.
Removed unused variable `prefix' in macros for generating c*r
functions and setf functions.  The prefix was used in order to
generate error messages that mention the name of the original
function, as opposed to relying on elementary car and cdr operations
to report errors.  But this idea still doesn't work, because the use
of a function can be the result of a macro expander, so we have to
find a more general mechanism.  Current thinking is that we should not
mention such detail at all, and instead trace the error back to the
source code that the user supplied, and show the original form.

As a result of this thinking, I removed the functions for generating
error messages for the c*r functions and setf functions.