Made the slot of the compilation conditions contain an ast instead of a form.
Changed the implementation of make-ast so that it can handle dotted
lists as well.

Added comment concerning how the environment code is wrong and how
it should be fixed.

Modified the fixup routine used after macro expansion so that it
can handle dotted lists.

Added comment about how macro expansion should be done.

Started implementing a function to parse an ordinary lambda list.