Introduced case-conversion tables for ascii characters for readtable cases :downcase and :preserve in addition to :upcase.
One of those case-conversion tables is now passed to the specialized
reader function (which was also renamed to reflect that).

The read function now dispatches on the readtable-case of the
readtable.  A readtable-case of :invert is still not handled.

Fixed an error whereby a parameter declared ignored was actually used.

Fixed two errors in copy-readtable that didn't copy the readtable-case.