Implemented subseq (function).
Implementd copy-seq.

Implemented elt.

Updated status.text to reflect the additions above.

Added subseq, copy-seq, elt, delete, delete-if, and delete-if-not to
shadowed symbols in packages.lisp.

Moved the invalid-bounding-index conditions closer to the beginning
of sequences.lisp because they are needed earlier.

Implemented all special versions of the count function.  Remains to
write the function itself, and then to derive count-if and
count-if-not from it.