Introduced a new condition `must-be-list' for situations where a cons or nil is required.
Modified the pop macro to check for its argument being a list and
signaling this condition otherwise.

Added a condition reporter for the new condition.

Removed the use of pop in other functions to avoid notes by the
compiler that indicated that the error situation checked for can never
occur (because those functions check the situation themselves).