Introduced a macro with-alist-elements that executes the body only if a cons is found, and which signals an error when an atom other than cons is found, and finally signals an error if the alist is not a proper list.
Used the above macro in the definition of the assoc* and rassoc*
family functions.

Replaced the direct use of specialized assoc function in sublis by
uses of the new macro.  This has two advantages: first, error messages
are given in terms of sublis (rather than assoc) and there is no need
to pass a test function with reversed arguments, because the call to
the test function is in the body of the macro, so just reverse them
there.