Modified macro with-proper-list-rests to evaluate the `name' argument.
Modified users of that macro to take the modification into account.

Introduced a new macro with-proper-list-elements similar to
with-proper-list-rests except that it loops over the elements
of the list instead of the rests.

Made use of the new macro in specialized versions of the union
and intersection functions.

Modified the specialized member functions to take a `name' parameter.
This parameter allows other functions to use these specialized member
functions, while still preserving the property that conditions are
reported in terms of the standard function that was called directly
by user code.

Modified the specialized union functions to use appropriate
specialized member functions instead of explicitly looping over
the second list.

All tests pass.