Renamed the specialized versions of the member function to conform to recent naming standards.
Introduced four new specialized version of the member function where
the test takes the arguments in reversed order.  This is needed in the
set-exclusive-or function when the elements of the second list are
tested for membership in the first list.

Modified special versions of the set-exclusive-or function to use new
specialized versions of the member function in order to signal more
specific error conditions when one of the lists is not a proper list.

All tests pass.