Modified the setf expander for getf so that the optional variable is ignored.  This avoids a warning message about an unused variable.
Fixed a problem in the remf macro that would have made
error reporting fail in case the list weren't a property list.

Changed the pushnew macro so that if it detects (at compile time) both
a :test and :test-not, it no longer signals an error (which would
terminate compilation) but instead it signals a warning, and generates
code to signal an error when the generated code is executed.