Defined a new condition used by the setf expander and the setf function for nth so that a precise error message can be given when there are not enough top-level CONS cells of the list.
Replaced the use of define-setf-expander by the long form of defsetf
for the setf expander of rest.

Defined the function (setf rest) so that it can be used with funcall
and apply.  Added (fmakunbound '(setf rest)) so as to avoid a
compilation warning when the setf expander for rest is defined.

Modified the setf function for nth so that it counts the number of
CONS cells, and keeps the original list for better error reporting.

Added a setf expander for nth.  Added (fmakunbound '(setf nth)) to
avoid compilation warning when the setf expander is defined.

Added comment headers for copy-list, list-length, make-list, nthcdr,
nth, and copy-tree.

Added a comment with respect to copy-tree indicating that it could
probably be done better by iterating on the CDR and recursing on the
CAR.

Removed several comments preceding special versions of functions,
describing what the specialization is.  These are no longer necessary
because the recent naming convention for such functions makes this
information explicit.

Added a (US English) condition reporter for the new condition
described above.

Updated status.text to reflect the additions to (setf nth).