More conditions.  Most of them copied from the cons-high module.
Added name-mixin to existing conditions so that the source of the
problem will be reported.

Improved skip-to-start to check for proper lists, and to take a
NAME parameter allowing us to signal the origin of the problem.

Used skip-to-start instead of nthcdr in many functions to skip a
prefix of a list.  This modification makes it possible to check that
the start index is a valid sequence index, and that the list is a
proper list.

Added function tail-must-be-proper-list used to check the tail
of a list when traversal reaches the end.

Modified many functions to use

         (loop for remaining on ...
               for element = (car remaining)
               ...
               finally (tail-must-be-proper-list ...))

instead of

         (loop for element in ...
               ...)

So that we don't leave it up to endp to signal that a list is not
a proper list.

Continued work on migrating to recent naming conventions by replacing
"unbounded" by "end=nil", "bounded" by "end=other", "key" by
"key=other", "identity" by "key=identity" as well as some other minor
improvements.