New method for traversing lists from the end, which is required by some sequence functions, in particular count when from-end is true. This method uses recursion, and traverses during backtrack, but it limits the stack depth to some fixed amount that can be set for each implementation.
Modified existing specialized versions of count to take advantage of
the new method.  Some of them were correct in that they traversed from
end, using an auxiliary vector.  Some of them were wrong, because they
would scan from start to end even though key=other, so the key
function could have side effects, and so scanning from start to end
would have been a violation of the standard.

This new method gives us a very fast version of count when from-end is
true and either key is not identity or either test or test-not is
neither eql nor eq, i.e., when the test and/or the key might have side
effects.  We have observed execution times of only 1/3 of those of
SBCL on long lists.

Even though there is no requirement that the list be scanned from the
end when from-end is true in functions such as find, scanning from the
beginning can be penalizing when the test and/or the key functions are
costly to execute.  We might use our new method for those cases as
well, thus potentially saving a lot when the item looked for is close
to the end of the list.

More utility functions to use at the end of lists.

More renamings in order that the code conform to recent naming
conventions.

The entire code is in a limbo now.  Many tests fail.  This is work in
progress to get it back into a reasonable state.