Introduced a utility function named copy-prefix to use with the remove-family functions when seq-type=list.  Also applied some code factoring to use other utility functions.
Currently, the find-family functions are all wrong for from-end=true
and count=other, because they destructively reverse the interval to be
checked.  It does put it back later, but in the presence of threads,
the current behavior is probably not wanted.  The plan, instead, is to
use the new method for scanning lists from end without consuming too
much stack.