Changed from-end=t to from-end=true everywhere so as to reflect the fact that from-end can be any generalized boolean.
For count, count-if, and count-if-not: introduced special versions
that are specialized on all the keyword parameters, but not on the
sequence type, and changed the main functions to use those new
versions.  The reason for this is that the corresponding compiler
macros (that have not been written yet) will not be able to determine
the sequence type, but can determine they keyword parameters, so the
compiler macros will use these new specialized functions.  The result
is that only the sequence type will have to be determined at execution
time, or perhaps not even then if the compiler can determine the type
by using type inference.