Factored out description of FIND, FIND-IF, and FIND-IF-NOT.
Factored out description of COUNT, COUNT-IF, and COUNT-IF-NOT.

Added a description of the PREDICATE SEQUENCE pair of arguments.

Renamed the description of satisfying the test to reflect that
it is a two-argument test.

Added descriptions of satisfying one argument tests, a positive test
and a negative test.

Added docstrings for FIND-IF, FIND-IF-NOT.

Added docstrings for COUNT-IF, COUNT-IF-NOT.