Introduced the concept of "host function".
This is the great cheat object.  We define a target object whose linkage
code calls a new machine instruction.  This new machine instruction
simulates a function call in the target by calling a function in the
host, using the values from the argument registers, and returning values
there as well.