Removed accessor generic functions for timestamps.
The plan is to detect obsolete objects by using the unique class
number rather than using timestamps.  The unique number of the class
will be stored in the instance just like the timestamp used to be, but
it will not be available as an explicit slot, so no replacement
accessors are required.