Removed timestamp as a slot in STANDARD-OBJECT.
We no longer use timestamps, and instead we store the unique number of
the class in the instance.  However, the class unique number will
always be the first element of the contents vector so we do not need
to turn it into an ordinary slot.  For that reason, the slot
specification is removed entirely from the DEFCLASS form.