Removed slot INSTANCE-SLOTS from STANDARD-OBJECT.
We store a copy of the effective slots of the class in the instance of
STANDARD-OBJECT as before, but we no longer treat the place where this
information is stored as an ordinary slot.  Instead, in a standard
object, it will be stored at offset 1, right after the unique number
of the class.