Use a special function as value in function cell to mean funbound.
Instead of using the ordinary immediate unbound value, for function
cells, we use a particular function instead.  That function takes an
arbitrary number of arguments and it signals an error.  This way, when
a call is made to a named function in the ordinary way, there is no
need for an explicit check to see whether the function name is fbound
or not.  Should it not be fbound, the special function signals an
error.