Split the support code for COMPUTE-EFFECTIVE-METHOD into 3 files.
Before, there were two ordinary functions representing versions of the
default action of COMPUTE-EFFECTIVE-METHOD in the same file.  The two
functions were called COMPUTE-EFFECTIVE-METHOD-DEFAULT-1 and
COMPUTE-EFFECTIVE-METHOD-DEFAULT-2.  That solution was not great
because then the client source code would have to be edited in order
to choose one or the other.

The new solution is to keep common support code in the original file
and to make two additional files, each containing a function named
just COMPUTE-EFFECTIVE-METHOD-DEFAULT.  Choosing one or the other is
now a question of choosing which of the two new files to use in the
build process.