Defined additional reader M-SPECIALIZERS similar to METHOD-SPECIALIZERS.
During phase 3 of bootstrapping, the generic function
METHOD-SPECIALIZERS is used to compute wich methods are applicable on
a bridge generic function, so that the entire mechanism for executing
generic functions can work.  However, at the same time,
ADD-DIRECT-METHOD is called by ADD-METHOD for each specializer of a
method, so there must be a way to access the specializers of a target
method as well.  So we have a second reader for the slot containing
the specializers of a method which is used for target generic
functions.