Extracted two versions of COMPUTE-DISCRIMINATING-FUNCTION-DEFAULT.
The first version works by closing over the GENERIC-FUNCTION argument
and returning a closure that has the signature of the discriminating
function and that calls DEFAULT-DISCRIMINATING-FUNCTION with the
closed-over generic function and the arguments.  This version could be
used in an initial image that does not have the compiler present yet.

The second version works by compiling in the GENERIC-FUNCTION argument
as a constant, creating a function that again has the signature of the
discriminating function and that calls DEFAULT-DISCRIMINATING-FUNCTION
with the compiled-in constant and the arguments.  This version
generates a discriminating function that is probably a bit faster than
the previous one.  The main use for it though, is that it can be used
with the cross compiler to generate a target function for a generic
function.