Skeleton version of COMPILE in phase 3.
Right now, this file only contains proof of concept.  It redefines
COMPILE to call ALLOCATE-INSTANCE of the built-in class FUNCTION.  The
first improvement will be to call MAKE-INSTANCE instead so that the
instance gets properly initialized.  Once the cross compiler works, it
will call the cross compiler to generate a target function.