Switched to using compiled version of accessor methods.
Before, the method-function of an accessor method was created as a
closure, closing over the slot name and then calling SLOT-VALUE or
(SETF SLOT-VALUE).  This is fine in a running image where it is
possible to execute code to create the closure.  However, this method
does not work when the cross compiler is used to create code for the
target, because the cross compiler can not execute target functions.

Now, during bootstrapping, the method-function of an accessor is
created with the slot name compiled in as a constant.  This method
does not require target code to be executed in order to create a
target function, so it is better adapted for use with the cross
compiler.

During phase 2 of the bootstrapping process, it does not matter which
version we use, but during phase 3, it is important to use the
compiled version.  Since it does not matter in phase 2, we might as
well use the same version there.