Use unique class numbers in generic function dispatch.
This version of the generic function dispatch is a step on the way
toward the final goal of fast generic function dispatch.

So far, we have implemented the CALL HISTORY which is basically a
cache of what COMPUTE-APPLICABLE-METHODS-USING-CLASSES returns,
together with the resulting effective method.  The discriminating
function consults the cache before calling the machinery of computing
applicable methods and effective methods.