Change the definition of BRIDGE-GENERIC-FUNCTION.
Before, BRIDGE-GENERIC-FUNCTION was a subclass of the host system
class named CL:STANDARD-GENERIC-FUNCTION and of the host class named
SICL-CLOS:STANDARD-GENERIC-FUNCTION.  This implementation created a
problem because initargs that are keywords would be common for both
those superclasses which created conflicts.

Now, BRIDGE-GENERIC-FUNCTION is a subclass of the host system class
named <host-clos-package>:FUNCALLABLE-STANDARD-OBJECT and again of
the host class named SICL-CLOS:STANDARD-GENERIC-FUNCTION.

Furthermore, when dealing with bridge generic functions, the function
SICL-CLOS:SET-FUNCALLABLE-INSTANCE-FUNCTION now calls the host
function with that name.

The main disadvantage of the new solution is that the we now need
elements that are not in the COMMON-LISP package, so we become
dependent on the particular host system used for bootstrapping.