Replace reader and writer methods by methods doing direct access.
Instead of making an effective method directly out of the methods
returned by COMPUTE-APPLICABLE-METHODS-USING-CLASSES, we make an
additional pass over the methods, and if we see a standard reader
method or a standard writer method, we replace it with a method that
uses STANDARD-INSTANCE-ACCESS to access the slot directly.

We can not make this replacement earlier because only at this point do
we know the exact class of the argument, and we need to know that
because the slot may have a different location in different classes.
Also, we still want to keep the original reader or writer method to
store it in the call history, so that if that method is removed, we
will find the call-history entry containing it, and delete the entry.