Set initial method function of accessor method to a dummy.
Before, we set the method function of an accessor method to a function
calling SLOT-VALUE.  There are several reasons we do not want to do
that.

The main reason is that SLOT-VALUE is slow and we want accessors to be
fast.  Therefore, when we compute an effective method from a list of
methods containing an accessor method, we replace that accessor method
with one that has a method function that does direct access using
standard-instance-access.  Since the accessor method is always
replaced, it does not matter what we set the method function to
initially.

The secondary reason is that during bootstrapping we do not want the
definition of a class to call SLOT-DEFINITION-NAME.  The reason for
not wanting such a call is that SLOT-DEFINITION-NAME is also used in a
different context, namely by SHARED-INITIALIZE.  When bridge classes
are instantiated, in the first context, SLOT-DEFINITION-NAME would
access a slot in a host instance, and in the second context, it would
access a lot in a bridge instance.  They can not both exist
simultaneously.