Do not make the default discriminating function call SPECIALIZER-PROFILE.
At the end of phase 3 of the bootstrapping process,
SPECIALIZER-PROFILE needs to be an installed bridge generic function.
For that reason, the default discriminating function on bridge generic
functions can not be allowed to call it, because for that to work,
SPECIALIZER-PROFILE needs to be a host generic function.

The problem is solved by passing the specializer profile as an
argument to the default discriminating function by the installed
discriminating function.  This works because the discriminating
function is recomputed whenever the specializer profile is modified.