Define new accessor PRECEDENCE-LIST, and remove writer C-PRECEDENCE-LIST.
During class finalization, we need to store the precedence list
computed with COMPUTE-CLASS-PRECEDENCE-LIST.  The AMOP says that step
1 of finalization does that and then associates the result with the
class.  The AMOP also hints that steps 2 and 3 of finalization might
use CLASS-PRECEDENCE-LIST to access the result of step 1.  However,
this is not possible, because the AMOP also says that
CLASS-PRECEDENCE-LIST signals an error if the class is not finalized,
and during finalization, this is not yet the case.  Our solution is to
introduce an accessor (whose name is not exported) called
PRECEDENCE-LIST so that we have both an alternative reader for the
slot and a writer to store the result of calling
COMPUTE-CLASS-PRECEDENCE-LIST.