Excluded component accessor-defgenerics from phase 1.
This is the beginning of an attempt to eliminate this file all
together.  Its sole purpose at the moment is to facilitate the
implementation of ENSURE-GENERIC-FUNCTION and DEFGENERIC later in the
bootstrapping process, and to avoid compiler warnings in phase 1.  But
it is redundant, so it creates a maintenance problem.  The goal is to
have a single definition of DEFGENERIC, and to redefine
ENSURE-GENERIC-FUNCTION as needed.