Rearranged class definitions.
It is important to order the class definitions in this file
depth-first pre-order, because that is how the classes will eventually
be finalized.  Since finalization is where unique class numbers are
assigned, this order ensures that a subtree of classes have
consecutive class numbers, thus making dispatch faster.

The result of this rearrangement is a depth-first pre-order as before,
except that now, the definition of BUILT-IN-CLASS comes AFTER the
definition of the subtree with REGULAR-CLASS as its root.  The reason
for this rearrangement is that the subtree with REGULAR-CLASS as its
root has some accessors for which we have DEFMETHOD forms spacialized
to BUILT-IN-CLASS, and this rearrangement ensures that the
corresponding generic functions are created when the DEFMETHOD forms
are seen, thus avoiding warnings from the compiler.