Redefine MAKE-INSTANCE in phase 4.
In phase 4, we can not instantiate target classes.  The reason is that
some slots may have initfunctions on them, and those initfunctions are
target instances that we can not execute here.

Instead, we instantiate the corresponding bridge classes, and then
immediately patch the instance we obtain, so that its class is a
target class, and (if it is a standard instance) also its list of
effective slots that were copied from its class so that it is a list
of target instances.