Beginning of a cleaner and more general bootstrapping process.
The previous attempt at bootstrapping CLOS basically works, and this
process will use most of the ideas.  The previous one got messy,
though, because it used a single package, requiring the symbols to
change roles between bootstrapping phases.  We hope to make things
simpler and cleaner with the new process.

In addition, we want to include bootstrapping of the rest of the
system here as well; not only CLOS.  The purpose of the first few
phases is still to get the MOP hierarchy in place, though.