Removed symbolic links to GFs and methods for slot definition classes.
The methods on the generic functions DIRECT-SLOT-DEFINITION-CLASS and
EFFECTIVE-SLOT-DEFINITION-CLASS are specialized to REGULAR-CLASS which
is a common superclass of the two classes FUNCALLABLE-STANDARD-CLASS
and STANDARD-CLASS.  However, during bootstrapping, we these functions
are also called with built-in classes as arguments.  For that reason, we
need to use a more permissive version of these two functions.