Pass function INITIALIZE-INSTANCE to default action of MAKE-INSTANCE.
The previous version of MAKE-INSTANCE-DEFAULT called functions that
apply to the class, and also INITIALIZE-INSTANCE which applies to the
instance.  During bootstrapping, the functions that apply to the class
need to be host generic functions, and initialize-instance needs to be
a bridge generic function.  However, (again during bootstrapping)
names of host generic functions and names of bridge generic functions
live in different packages, so we can not mix them.  The solution was
for MAKE-INSTANCE to pass the bridge function to MAKE-INSTANCE-DEFAULT
and to let MAKE-INSTANCE-DEFAULT refer to host functions directly.