Decrease coupling between the two phases.
This is the start of an effort with the ultimate purpose of making the
names of both bridge classes and target classes be symbols in the
correct package, i.e. some of the names will be in the COMMON-LISP
package (such as STANDARD-CLASS, STANDARD-OBJECT, FUNCTION, etc) and
the remaining ones will be in the SICL-CLOS package.