Pass INITIALIZE-BUILT-IN-INSTANCE as argument to MAKE-BUILT-IN-INSTANCE-DEFAULT.
This modification is similar to the one recently done for
INITIALIZE-INSTANCE and MAKE-INSTANCE-DEFAULT.  Again, the reason is
that MAKE-BUILT-IN-INSTANCE-DEFAULT does two things that are different
in nature, and which influences how we do the bootstrapping.  First it
calls ALLOCATE-BUILT-IN-INSTANCE which uses properties of the CLASS to
be instantiated.  Then it calls INITIALIZE-BUILT-IN-INSTANCE which has
more to do with the INSTANCE than with the class.