Moved the check to see whether the class is finalized to MAKE-INSTANCE.
The AMOP says that ALLOCATE-INSTANCE checks whether the class is
finalized, and if not, calls FINALIZE-INHERITANCE.  However, the
INITARGS received by ALLOCATE-INSTANCE should be the defaulted
initargs, and computing the defaulted initargs requires the class to
be finalized.  I peek at PCL shows that the class is finalized in
MAKE-INSTANCE, before ALLOCATE-INSTANCE is called, which makes more
sense.