Added slot INVOCATION to the CONTEXT class.
Here is why:

I define the concept of OWNERSHIP of instructions.  The reason for
this concept was to detect a control transfer from an instruction with
one owner to an instruction with a different owner, which would then
be a non-local control transfer.  However, I though I could define
ownership as being reachable from transitions starting at the
ENTER-INSTRUCTION.  As it turns out, this is not possible because
there can be a non-local control transfer to an instruction that is
not reachable from its owner.  As a result, I needed a different way
of determining whether a control transfer is local or not.  What I
will ultimately do is to mark BLOCKs and TAGBODYs by the INVOCATION
which is ether the ENTER-INSTRUCTION to which they belong, or NIL if
they are outside of any ENTER-INSTRUCTION.  When a RETURN-FROM or a GO
is detected where the target has a different invocation from the one
in the current context, it is considered a non-local control transfer.