Add another parameter STATIC-ENVIRONMENT.
This commit is the first in a series with the purpose of making
TRANSLATE-HIR accept more HIR instructions.  In particular, the
instructions on CELLs that are introduced by the escape analysis must
be handled.  For that to work, we need to manage a static environment
that maps indices of cells to symbols that stand for host variables.