Parse (values ...) types at AST level
THE-AST and THE-VALUES-INSTRUCTION now have accessors for required,
optional, and rest components of their types, instead of a blob.
(THE-INSTRUCTION is unchanged.)

values types involving &optional and/or &rest are almost completely
unused (not counting sbcl's &optional, which is for special semantics)
but complicate everything enormously. The worst kind of feature.

This at least makes them a bit more tolerable, by letting you not
reparse everything four times as you try to do everything completely
scrupulously.