type inference: detect compute-initial-dictionary problem and signal
Currently liveness is broken such that type inference will return an
empty dictionary for certain kinds of loops. That will be difficult to
fix, but in the meantime, find-type should not treat "no information" as
"it's of type NIL" as it did before it commit: it should signal an error
saying that type inference did not work.