cleavir environment: refactor optimize information for policies
Rather than storing individual qualities, the whole thing is stored at
once, along with a complete policy.

The basic goal here is to make sure policies don't have to be
recomputed. Because every AST and instruction has a complete policy in a
slot, recomputing it for every AST would be very wasteful.

This imposes a subtle requirement on implementations: the global
environment should cache the global policy, and the implementation
should recompute it every time it sees a proclaim optimize.