Let HIR know about DYNAMIC-EXTENT declarations
This adds dynamic-allocation -ast and -instruction. They have comments.
Basically they're like THE, they have no effect but tell analysis that
their input is dynamic-extent.

Because of how dynamic-extent works I felt the best thing to do
in generate-AST was to mark the value of the binding as such, and not
uses. This makes the convert-special for let, flet, etc. even uglier.

These instructions aren't used yet.