Rewrite Kildall type inference with better descriptors
Long version is in Descriptors/lattice-descriptor.lisp.

Implementation status: types inferred for my half-page numerical program
seem to be correct.

 * typeq elimination still not done
 * Drawing newly broken (needs descriptor->specifier)
 * Newly required environment accessors are sbcl-only hacks

Profiler says 80% of time is in pool-meet at ~1 ms per call, but
object-meet is ~0 per call and just called a lot. So probably doing the
liveness stuff again should fix that, and I'm not going to worry about
performance until I do that.