Generate-AST: alter how special bindings work and add AST primop
Previously, special variable bindings resulted in a generated call to
CLEAVIR-PRIMOP:CALL-WITH-VARIABLE-BOUND. This was done very directly so
implementations could not choose to make C-W-V-B a macro or special
operator, or to put a compiler on it. Since the implementor has to
define this operator anyway, this restriction is unnecessary.

In the course of this I added a CLEAVIR-PRIMOP:AST special operator.
This is something I'd been talking about with beach that turns out to be
useful here. It just inserts a provided AST directly into the AST being
compiled. E.g.

(+ (cleavir-primop:ast #<IMMEDIATE-AST 4>) 7) compiles like (+ 4 7).