Revert "cleavir segregate-lexicals: Allocate cells as late as possible"
This reverts commit 0f94365b911d2c26086088353f24e95e0b21c4a2.

Clasp seems to have a write to an uninitialized cell somewhere with
this commit, and it seems likely to be flawed. map-instructions doesn't
handle control flow in quite the right way.

beach has another plan for fixing this involving changing how LET and
so on are compiled, which will avoid this problem.