Cleavir liveness: switch back to old implementation
I rewrote liveness with bit vectors. Testing this in clasp had it working, but
less efficient. Given that efficiency was the only reason for the rewrite this
was rather disheartening.

Particularly: using bit vectors means that no allocation has to be done in the
main Kildall loop. That's cool, but it means that each instruction is associated
with a bit vector (its pool) that has one bit for every variable in the graph.
This means that the bulk of the space required is the number of instructions
times the number of variables.

This is contrasted with the old implementation that just collected lists of
variables. Overall the bit vector implementation trades space for speed. I
expected that the use of bits instead of object references would make the
space usage tolerable, and perhaps even result in an overall gain, but I was
mistaken.

A few basic tests with Clasp showed that the slow-to-compile code would have a
few hundred instructions (normal) and tens of thousands of variables (eh). This
easily results in hundreds of KB of storage, if not megabytes. Importantly,
large parts had to be accessed within the main loop all at once. Probably
defeated processor caching.

This effect was obvious and severe. The bit vector implementation was not just a
little slower. I thought that the compiler hung.

It's possible that bit vectors could actually be more efficient, but only if
there are far less variables (in normal code), and even then it would still get
bad on code with a lot of variables. The fact is that at any given instruction,
there are almost always less than a dozen or two live variables, so a bit vector
of all variables will be very sparse.

There are possible improvements to be made with this implementation. Sets as
lists is not great, lots of O(n^2). But it might not be worth improving, at
least not right off.